+++ src/visualizers/highcharts/charts/scatter.ts
@@ -1,0 +1,21 @@
+'use strict';
+
+import { HighchartsChart } from './highchartsChart';
+import { ChartTypeOptions, UNSTACKED } from '../chartTypeOptions';
+
+export class Scatter extends HighchartsChart {
+    //#region Methods override
+
+    protected getChartTypeOptions(): ChartTypeOptions {
+        return {
+            chartType: 'scatter',
+            plotOptions: {
+                scatter:  {
+                    stacking: UNSTACKED
+                }
+            }
+        }
+    }
+
+    //#endregion Methods override
+}
